+++ database/migrations/2014_10_12_100000_create_password_resets_table.php
@@ -27,6 +27,8 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('password_resets');
+        if (!app()->isProduction()) {
+            Schema::dropIfExists('password_resets');
+        }
     }
 };
