+++ database/migrations/2019_12_14_000001_create_personal_access_tokens_table.php
@@ -32,6 +32,8 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('personal_access_tokens');
+        if (!app()->isProduction()) {
+            Schema::dropIfExists('personal_access_tokens');
+        }
     }
 };
